refactor : column, value clause 로직에서 공통 부분 메소드 추출
